Fixed generation of EventInfo from SQL query

+++ src/main/java/team/gif/gearscout/matches/EventInfo.java
@@ -4,5 +4,5 @@ public record EventInfo(
 	Integer gameYear,
 	String eventCode,
 	String secretCode,
-	Integer matchCount
+	Long matchCount
 ) {}
